ml-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
python-dotenv>=1.0.1
Pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.15.0
google-generativeai>=0.3.0
